refactor(ui/web): migrate ExchangeLists from @starboard
