+++ frontend/src/pages/Dashboard/src/components/ExchangeLists/components/TradeHistoryCard.tsx
